refactor: migrate to isar v4.0

+++ lib/lib/core/storage/messages.storage.dart
@@ -1,39 +1,39 @@
+import 'package:barbox/core/storage/isar/base.db.dart';
 import 'package:barbox/core/storage/isar/local_account.db.dart';
+import 'package:barbox/types/messages/message.dart';
 import 'package:injectable/injectable.dart';
 import 'package:isar/isar.dart';
-import 'package:barbox/core/storage/isar/base.db.dart';
-import 'package:barbox/types/messages/message.dart';
 
 @LazySingleton()
 class MessagesStorage {
   Future<void> saveMessage(Message message) async {
-    await isarInstance.writeTxn(() async {
-      await isarInstance.messages.put(message);
+    await isarInstance.writeAsync((isar) async {
+      isar.messages.put(message);
     });
   }
 
   Future<void> saveMessages(List<Message> message) async {
-    await isarInstance.writeTxn(() async {
-      await isarInstance.messages.putAll(message);
+    await isarInstance.writeAsync((isar) async {
+      isar.messages.putAll(message);
     });
   }
 
   //Isar's put() method will either insert or update the object depending on
   //whether it already exists in the collection.
   Future<void> updateMessageSeen(String msgId, bool seen) async {
-    await isarInstance.writeTxn(() async {
-      Message? message = await isarInstance.messages.getById(msgId);
+    await isarInstance.writeAsync((isar) async {
+      Message? message =
+          await isar.messages.where().idEqualTo(msgId).findFirstAsync();
 
       if (message != null) {
         message = message.copyWith(seen: seen);
-        await isarInstance.messages.putById(message);
+        isar.messages.put(message);
       }
     });
   }
 
   Future<bool> containsMessage(Message message) async {
-    final entries =
-        await isarInstance.messages.where().idEqualTo(message.id).count();
+    final entries = isarInstance.messages.where().idEqualTo(message.id).count();
 
     if (entries > 0) {
       return true;
@@ -47,23 +47,20 @@ class MessagesStorage {
   }
 
   Future<void> deleteMessage(int isarId) async {
-    return isarInstance.writeTxn(() async {
-      await isarInstance.messages.delete(isarId);
+    return isarInstance.writeAsync((isar) async {
+      isar.messages.delete(isarId);
     });
   }
 
   Future<void> clear() async {
-    return isarInstance.writeTxn(() async {
-      await isarInstance.messages.clear();
+    return isarInstance.writeAsync((isar) async {
+      isar.messages.clear();
     });
   }
 
   Future<void> clearMessagesByAddress(LocalAccount account) async {
-    return isarInstance.writeTxn(() async {
-      await isarInstance.messages
-          .where()
-          .accountIdEqualTo(account.accountId!)
-          .deleteAll();
+    return isarInstance.writeAsync((isar) async {
+      isar.messages.where().accountIdEqualTo(account.accountId).deleteAll();
     });
   }
 }
